weather-bot/requirements.txt
python-telegram-bot==21.4
requests==2.32.3